+++ releaf-core/app/controllers/releaf/base_controller.rb
@@ -136,6 +136,16 @@ module Releaf
       end
     end
 
+    def toolbox
+      @resource = resource_class.find(params[:id])
+
+      respond_to do |format|
+        format.html do
+          render partial: 'toolbox.items', locals: { resource: @resource }
+        end
+      end
+    end
+
     def destroy
       check_feature(:destroy)
       @resource = resource_class.find(params[:id])
@@ -474,9 +484,7 @@ module Releaf
         create:            true,
         destroy:           true,
         index:             true,
-        # enable toolbox for each table row
-        # it can be unnecessary for read only report like indexes
-        index_row_toolbox: true
+        toolbox:           true
       }
       @panel_layout      = true
       @resources_per_page    = 40