+++ releaf-core/app/middleware/releaf/routes_reloader.rb
@@ -16,7 +16,7 @@ module Releaf
 
     def self.reload_if_expired
       # TODO Node class should be configurable
-      return unless ::Node.updated_at.present? && @updated_at < ::Node.updated_at
+      return unless ::Node.updated_at.present? && @updated_at && @updated_at < ::Node.updated_at
       Rails.application.reload_routes!
       routes_loaded
     end